row/requirements.txt
Django==5.2.1
psycopg2-binary==2.9.9
pillow==10.3.0
crispy-tailwind==1.0.3
djangorestframework==3.15.2
django-environ==0.11.2
drf-spectacular==0.28.0
django-extensions==4.1
matplotlib==3.10.3